Fixes from @ebaker PR comments.
