+++ .eslintrc.js
@@ -46,6 +46,7 @@ module.exports = {
         allowTaggedTemplates: true,
       },
     ],
+    "no-console": 1,
     "comma-dangle": 0,
     "jsx-quotes": [2, "prefer-double"],
     "linebreak-style": ["error", "unix"],
