+++ packages/example/public/keplr-extension.png
